Modify the way the home page is rendered
I've added back the two-column layout from the laravel rewrite, but due
to some issues getting the columns to work directly in the markdown,
only the left hand column is defined in _index.md. This also means I
can't use the youtube shortcode for the introduction video, so I've used
bootstrap's embed classes and YouTube's own iframe to achieve the same
thing. The "Latest News" section is also now pure html, which is
annoying, although use of hugo's 'markdownify' function may be able to
fix that.