Make the Blog a section instead of a category
This just fits in better with the new (and to be honest even the old) layout
with sub-categories. The blog should be a leaf nnode: it shouldn't have
sub-categories in it, so I made it a section to achieve this.

This does make the URLs for the navbar pages inconsistent, but I don't consider
this a big deal, nor can I think of a non-hacky way to fix it.